Overhaul post clone/sync handling
When calling sync_posts, the context (sync or clone) is passed, dictating what rules are fetched if not provided, and how to treat preparing the rules.

When syncing, post_status is allowed, but not for cloning, whereas post_content is allowed for cloning.

So now, when cloning a post, only the bare minimum of title and type are copied (and title created), while the rest of the fields are actually cloned according to the rules dictacted in the synchronizer settings.

TL;DR the post_field clone rules are actually respected now, and new translations for published posts stay pending.