Fix check_blog_for_support to test if plugin is literally active
If the plugin was previously installed on the site being switched to but deactivated, then the hooks would be restored when they shouldn't be.

Since is_plugin_active() is admin-only and would require loading the somewhat large wp-admin/includes/plugin.php file just to use a simple function, I created an nlingual-specific version that covers both local and network plugins in one go.

Also required adding a new NL_* constant for storing JUST the plugin slug, not the path.