Patched update notice printing
This hook is called with a single notice block with a single p tag, making separate notices for updates difficult. This will close the currently open notice elements before printing a separate on, which will persist after the inline update is completed. It will re-open a blank div to be closed instead, to avoid HTML nesting bugs.