Fix date sync; bypass date clearing on drafts
When updating a draft post, the GMT date is cleared unless set by the user (via edit_date), so add that to the changes list if syncing the GMT date.

Main use case: creating a clone of a published post causes publishing the translation to use the current rather than the original's.