rewrite unlocalized location handling
If a location is marked as supported in the past but then unmarked later, it may cause the original to be continually set to the value of the vestigial default language version due to the array_merge with existing settings. This now strips out localized entries for no-longer-localized nav menu locations.