Add blog language map; track current language between blogs
For use cases where switching to a blog that doesn’t use the current language.

Current solution: before switching, log the current language to a list by blog ID. When switching, find the counterpart to that language on this new blog, falling back to the default language. When switching back (or at least to one that’s already been switched to) use that cached language.

TODO: figure out how to solve possible issues when switching language within/between blog switches.